requirements.txt
asgiref==3.8.1
attrs==24.3.0
# backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.16
django-crispy-forms==1.14.0
django-filter==24.3
django-mail-templated==2.6.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.8
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.2
pkgutil_resolve_name==1.3.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
